Use uniform indentation and eliminate empty lines
Code that this properly indented is easier to read.  You don't have to
do this yourself, there are tools that do this for you.  Sublime Text
has a way to do this[1], but that won't work on Windows (only Linux/OS
X).

You can use the following website though before submitting your changes
for review: http://phpbeautifier.com/

[1]: https://github.com/akalongman/sublimetext-codeformatter